refactor: extract challenge-delegation cluster from state.py God-Object (#1271)
